requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.1
django-appconf==1.0.5
django-celery-email==3.0.0
django-ckeditor==6.5.1
django-countries==7.3.2
django-crispy-forms==1.14.0
django-embed-video==1.4.4
django-import-export==2.8.0
django-js-asset==2.0.0
django-redis==5.2.0
django-yubin==1.7.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
kombu==5.2.4
lockfile==0.12.2
mail-parser==3.15.0
MarkupPy==1.14
numpy==1.23.2
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
typing_extensions==4.3.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
django-tawkto==0.5.1
django-richtexteditor==0.0.1.dev3
pymysql==1.1.1
